code cleanup and added copyright header
