+++ app/src/test/java/com/cognitree/kronos/executor/handlers/MockSuccessTaskHandler.java
@@ -1,5 +1,23 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.cognitree.kronos.executor.handlers;
 
+import com.cognitree.kronos.executor.model.TaskResult;
 import com.cognitree.kronos.model.Task;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
@@ -24,8 +42,8 @@ public class MockSuccessTaskHandler implements TaskHandler {
     }
 
     @Override
-    public Task.TaskResult handle(Task task) {
+    public TaskResult handle(Task task) {
         tasks.add(getTaskId(task.getName(), task.getJob(), task.getNamespace()));
-        return Task.TaskResult.SUCCESS;
+        return TaskResult.SUCCESS;
     }
 }
