+++ app/src/test/java/com/cognitree/kronos/scheduler/TaskServiceTest.java
@@ -1,3 +1,20 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.cognitree.kronos.scheduler;
 
 import com.cognitree.kronos.ApplicationTest;
@@ -28,10 +45,10 @@ public class TaskServiceTest extends ApplicationTest {
         Thread.sleep(100);
 
         TaskService taskService = TaskService.getService();
-        final List<Task> cognitreeTasks = taskService.get(workflowTriggerOne.getNamespace());
-        Assert.assertEquals(3, cognitreeTasks.size());
-        final List<Task> kronosTasks = taskService.get(workflowTriggerTwo.getNamespace());
-        Assert.assertEquals(3, kronosTasks.size());
+        final List<Task> workflowOneTasks = taskService.get(workflowTriggerOne.getNamespace());
+        Assert.assertEquals(3, workflowOneTasks.size());
+        final List<Task> workflowTwoTasks = taskService.get(workflowTriggerTwo.getNamespace());
+        Assert.assertEquals(3, workflowTwoTasks.size());
     }
 
     @Test
