+++ app/src/test/java/com/cognitree/kronos/scheduler/WorkflowServiceTest.java
@@ -1,3 +1,20 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.cognitree.kronos.scheduler;
 
 import com.cognitree.kronos.ApplicationTest;
@@ -21,7 +38,6 @@ public class WorkflowServiceTest extends ApplicationTest {
         workflowService.add(workflow);
         Assert.fail();
     }
-
 
     @Test(expected = ValidationException.class)
     public void testAddInValidWorkflowMissingTasks() throws Exception {
