+++ common/src/main/java/com/cognitree/kronos/model/TaskUpdate.java
@@ -1,0 +1,94 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.cognitree.kronos.model;
+
+import com.cognitree.kronos.model.Task.Status;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
+
+public class TaskUpdate {
+    @JsonSerialize(as = TaskId.class)
+    @JsonDeserialize(as = TaskId.class)
+    private TaskId taskId;
+    private Status status;
+    private String statusMessage;
+    private Map<String, Object> context = new HashMap<>();
+
+    public TaskId getTaskId() {
+        return taskId;
+    }
+
+    public void setTaskId(TaskId taskId) {
+        this.taskId = taskId;
+    }
+
+    public Status getStatus() {
+        return status;
+    }
+
+    public void setStatus(Status status) {
+        this.status = status;
+    }
+
+    public String getStatusMessage() {
+        return statusMessage;
+    }
+
+    public void setStatusMessage(String statusMessage) {
+        this.statusMessage = statusMessage;
+    }
+
+    public Map<String, Object> getContext() {
+        return context;
+    }
+
+    public void setContext(Map<String, Object> context) {
+        this.context = context;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof TaskUpdate)) return false;
+        TaskUpdate that = (TaskUpdate) o;
+        return Objects.equals(taskId, that.taskId) &&
+                status == that.status &&
+                Objects.equals(statusMessage, that.statusMessage) &&
+                Objects.equals(context, that.context);
+    }
+
+    @Override
+    public int hashCode() {
+
+        return Objects.hash(taskId, status, statusMessage, context);
+    }
+
+    @Override
+    public String toString() {
+        return "TaskUpdate{" +
+                "taskId=" + taskId +
+                ", status=" + status +
+                ", statusMessage='" + statusMessage + '\'' +
+                ", context=" + context +
+                '}';
+    }
+}
