+++ executor/src/main/java/com/cognitree/kronos/executor/handlers/TaskHandler.java
@@ -17,8 +17,8 @@
 
 package com.cognitree.kronos.executor.handlers;
 
+import com.cognitree.kronos.executor.model.TaskResult;
 import com.cognitree.kronos.model.Task;
-import com.cognitree.kronos.model.Task.TaskResult;
 import com.cognitree.kronos.model.Workflow.WorkflowTask;
 import com.fasterxml.jackson.databind.node.ObjectNode;
