+++ extensions/spark-handler/pom.xml
@@ -27,7 +27,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>spark-handler</artifactId>
-    
+
     <properties>
         <http.core.version>4.4.10</http.core.version>
         <http.client.version>4.5.6</http.client.version>
