+++ scheduler/src/main/java/com/cognitree/kronos/scheduler/TaskSchedulerService.java
@@ -23,8 +23,8 @@ import com.cognitree.kronos.model.MutableTask;
 import com.cognitree.kronos.model.MutableTaskId;
 import com.cognitree.kronos.model.Task;
 import com.cognitree.kronos.model.Task.Status;
-import com.cognitree.kronos.model.Task.TaskUpdate;
 import com.cognitree.kronos.model.TaskId;
+import com.cognitree.kronos.model.TaskUpdate;
 import com.cognitree.kronos.queue.QueueConfig;
 import com.cognitree.kronos.queue.consumer.Consumer;
 import com.cognitree.kronos.queue.consumer.ConsumerConfig;
